+++ packages/core/src/di/injector_compatibility.ts
@@ -8,14 +8,39 @@
 
 import {Type} from '../interface/type';
 import {stringify} from '../util/stringify';
-
+import {getClosureSafeProperty} from '../util/property';
 import {InjectionToken} from './injection_token';
 import {Injector} from './injector';
 import {getInjectableDef, ɵɵInjectableDef} from './interface/defs';
+import {ValueProvider} from './interface/provider';
 import {InjectFlags} from './interface/injector';
 import {Inject, Optional, Self, SkipSelf} from './metadata';
 
 
+/**
+ * An InjectionToken that gets the current `Injector` for `createInjector()`-style injectors.
+ *
+ * Requesting this token instead of `Injector` allows `StaticInjector` to be tree-shaken from a
+ * project.
+ *
+ * @publicApi
+ */
+export const INJECTOR = new InjectionToken<Injector>(
+  'INJECTOR',
+  -1 as any  // `-1` is used by Ivy DI system as special value to recognize it as `Injector`.
+);
+
+const _THROW_IF_NOT_FOUND = new Object();
+export const THROW_IF_NOT_FOUND = _THROW_IF_NOT_FOUND;
+
+export const NG_TEMP_TOKEN_PATH = 'ngTempTokenPath';
+const NG_TOKEN_PATH = 'ngTokenPath';
+const NEW_LINE = /\n/gm;
+const NO_NEW_LINE = 'ɵ';
+export const SOURCE = '__source';
+
+export const USE_VALUE =
+  getClosureSafeProperty<ValueProvider>({provide: String, useValue: getClosureSafeProperty});
 
 /**
  * Current injector value used by `inject`.
@@ -163,4 +188,53 @@ export function injectArgs(types: (Type<any>| InjectionToken<any>| any[])[]): an
     }
   }
   return args;
+}
+
+
+export class NullInjector implements Injector {
+  get(token: any, notFoundValue: any = THROW_IF_NOT_FOUND): any {
+    if (notFoundValue === THROW_IF_NOT_FOUND) {
+      // Intentionally left behind: With dev tools open the debugger will stop here. There is no
+      // reason why correctly written application should cause this exception.
+      // TODO(misko): uncomment the next line once `ngDevMode` works with closure.
+      // if(ngDevMode) debugger;
+      const error = new Error(`NullInjectorError: No provider for ${stringify(token)}!`);
+      error.name = 'NullInjectorError';
+      throw error;
+    }
+    return notFoundValue;
+  }
+}
+
+
+export function catchInjectorError(
+  e: any, token: any, injectorErrorName: string, source: string | null): never {
+  const tokenPath: any[] = e[NG_TEMP_TOKEN_PATH];
+  if (token[SOURCE]) {
+    tokenPath.unshift(token[SOURCE]);
+  }
+  e.message = formatError('\n' + e.message, tokenPath, injectorErrorName, source);
+  e[NG_TOKEN_PATH] = tokenPath;
+  e[NG_TEMP_TOKEN_PATH] = null;
+  throw e;
+}
+
+export function formatError(
+  text: string, obj: any, injectorErrorName: string, source: string | null = null): string {
+  text = text && text.charAt(0) === '\n' && text.charAt(1) == NO_NEW_LINE ? text.substr(2) : text;
+  let context = stringify(obj);
+  if (obj instanceof Array) {
+    context = obj.map(stringify).join(' -> ');
+  } else if (typeof obj === 'object') {
+    let parts = <string[]>[];
+    for (let key in obj) {
+      if (obj.hasOwnProperty(key)) {
+        let value = obj[key];
+        parts.push(
+          key + ':' + (typeof value === 'string' ? JSON.stringify(value) : stringify(value)));
+      }
+    }
+    context = `{${parts.join(', ')}}`;
+  }
+  return `${injectorErrorName}${source ? '(' + source + ')' : ''}[${context}]: ${text.replace(NEW_LINE, '\n  ')}`;
 }
